WIP: fp32 state across chunks, single-rounded operands, fp32 u and output
Second step on top of the fp32 resident state. Measured on tapped
GLM-5.3-Flash layer-0 inputs (T=13824, real carried state), relative L2
against an fp32 recurrence, output / final state:

  shipped kernel                 0.0050 / 0.0118
  fp32 resident state            0.0047 / 0.0031
  this commit                    0.0033 / 0.0014
  FLA chunk kernel (Triton)      0.0035 / 0.0020

On layer 1 the shipped kernel's final state is 0.105 / 0.071 off on the
two tapped chunks; this commit gives 0.0008 / 0.0010 (FLA 0.0011 /
0.0016).

Kernel 1 forms k_decayed, q_decayed, k_inv and k_restored from the fp32
normalized value with one bf16 rounding each, and applies the fp32
attention scale. Kernel 2 forms u = (v - kS) * beta in fp32 before the
INV GEMM, accumulates both output GEMMs in one fp32 accumulator and
rounds once at the store, and moves an fp32 state between the TMA
landing buffer and the fragments directly, so a chunk boundary no longer
passes through the bf16 tile. The fragment machinery moves to function
scope for that: the load has to finish before the entry barrier, after
which the input pipeline reuses the landing buffer. Checkpoint exports
still go through the bf16 tile.

torch_ref mirrors the new arithmetic. The output can differ from the
reference by one bf16 ulp because the two output GEMMs share one MMA
accumulator whose summation order torch cannot reproduce, and elements
that two large terms cancel to below the fp32 resolution of the tensor's
largest values are ordering noise; test_fwd checks the output under those
two rules and the final state exactly. Both tests pass on a GB10.

Kernel time on a GB10 is unchanged at T=16128/4096, H=16/64. Register
spills grow to 96-200 bytes in the full-width instantiations under
__launch_bounds__(192, 2).

End to end on a 4-way TP GLM-5.3-Flash deployment (SM90 plan fix on all
ranks), prefill throughput is 5-6% above the Triton path from 5k to 50k
tokens. A tool-call retrieval probe, 32 runs per length over four rounds
with an identical prefill chunk pattern in both arms: this kernel 34k
30/32, 42k 23/32; the FLA Triton chunk path 32/32 and 19/32; the shipped
kernel 13/16 and 5/16 over two rounds.

+++ csrc/smxx/fwd_kernel1.cuh
@@ -428,6 +428,8 @@ __global__ void __launch_bounds__(NumThreads, 8) _flash_kda_fwd_prepare(
                 Tensor s_kr = local_tile(tile_kr, thr2_2d, make_coord(0, t));
                 Tensor s_ki = local_tile(tile_ki, thr2_2d, make_coord(0, t));
 
+                // Each operand is formed from the fp32 normalized value and
+                // rounded to bf16 once. A product of two bf16 values rounds twice.
                 Tensor r_qd = make_tensor_like<BF16>(s_qd);
                 Tensor r_kd = make_tensor_like<BF16>(s_kd);
                 float q_inv = shared_storage.q_norm_inv.begin()[row];
@@ -435,11 +437,11 @@ __global__ void __launch_bounds__(NumThreads, 8) _flash_kda_fwd_prepare(
                 #pragma unroll
                 for (int v = 0; v < 2; ++v) {
                     float g = reg_g[pass][tile_idx][v];
-                    BF16 q = BF16(bf16_to_f32(reg_q[pass][tile_idx][v]) * q_inv);
-                    BF16 k = row < actual_len ? BF16(bf16_to_f32(reg_k[pass][tile_idx][v]) * k_inv_scale) : BF16(0);
-                    BF16 exp_cumsum = BF16(ex2_approx_ftz_f32(g));
-                    r_qd(0, v) = q * exp_cumsum * BF16(scale);
-                    r_kd(0, v) = k * exp_cumsum;
+                    float q = bf16_to_f32(reg_q[pass][tile_idx][v]) * q_inv;
+                    float k = row < actual_len ? bf16_to_f32(reg_k[pass][tile_idx][v]) * k_inv_scale : 0.f;
+                    float exp_cumsum = ex2_approx_ftz_f32(g);
+                    r_qd(0, v) = BF16(q * exp_cumsum * scale);
+                    r_kd(0, v) = BF16(k * exp_cumsum);
                 }
                 cute::copy(AutoVectorizingCopy{}, r_qd, s_qd);
                 cute::copy(AutoVectorizingCopy{}, r_kd, s_kd);
@@ -449,10 +451,10 @@ __global__ void __launch_bounds__(NumThreads, 8) _flash_kda_fwd_prepare(
                 #pragma unroll
                 for (int v = 0; v < 2; ++v) {
                     float g = reg_g[pass][tile_idx][v];
-                    BF16 k = row < actual_len ? BF16(bf16_to_f32(reg_k[pass][tile_idx][v]) * k_inv_scale) : BF16(0);
-                    BF16 inv_cumsum = BF16(ex2_approx_ftz_f32(-g));
-                    r_ki(0, v) = k * inv_cumsum;
-                    r_kr(0, v) = k * inv_cumsum * BF16(reg_gt[pass][tile_idx][v]);
+                    float k = row < actual_len ? bf16_to_f32(reg_k[pass][tile_idx][v]) * k_inv_scale : 0.f;
+                    float inv_cumsum = ex2_approx_ftz_f32(-g);
+                    r_ki(0, v) = BF16(k * inv_cumsum);
+                    r_kr(0, v) = BF16(k * inv_cumsum * reg_gt[pass][tile_idx][v]);
                 }
                 cute::copy(AutoVectorizingCopy{}, r_ki, s_ki);
                 cute::copy(AutoVectorizingCopy{}, r_kr, s_kr);
